Debuggato la generazione dei task creando un costruttore dei i container che fa da "copia"

+++ PureEdgeSim/com/mechalikh/pureedgesim/taskgenerator/Container.java
@@ -148,7 +148,7 @@ public interface Container extends QueueElement {
 	 * @param requestSize the file size of the task request to set in bits
 	 * @return the updated Task object
 	 */
-	double getFileSizeInBits();
+	long getFileSizeInBits();
 
 	/**
 	 *
